(1)  Added highlighting of the "sticky day" to the mini-cals, on top of highlighting the current day/week/month.
(2)  Implemented the "sticky day" in the week links, preserving the day of the week when moving to a week for which the current sticky day is invalid.

(3)  Added links to the View Day/Week/Month labels in the classic trailer, to bring it into line with the simple trailer and to assist navigation between the day/week/month views now that the "sticky date" concept is working.


git-svn-id: https://mrbs.svn.sourceforge.net/svnroot/mrbs/mrbs/trunk@918 1c18d0e4-b14a-0410-8b7b-fc87120d1eea