Temporary hack: make the whole cell clickable in the month view
When the table cells are very high (because there are many activities in
one day), then the div containing the link is not as large as the table
cell. Apparently, it is not very easy to do this using CSS. Therefoe, make
the table clickable via JavaScript. This is not ideal, because the user
has no visual cue that the cell is clickable. But still better than
nothing.